requirements.txt
# generic
clean-py==0.5
tqdm==4.64.0
pandas==1.4.3

# podcast scraping
pydub==0.25.1
beautifulsoup4==4.11.1
ftfy==6.1.1

# ASR
transformers==4.21.1
nemo-toolkit==1.10.0
pyannote.audio==2.0.1
pyannote.pipeline==2.3
numpy==1.22
nvsmi==0.4.2
jiwer==2.3.0

# evaluation
google-cloud==0.34.0
google-cloud-speech==2.15.1
google-cloud-storage==2.5.0
aiobotocore==2.4.0
boto3==1.24.66
botocore==1.27.66
s3fs==2022.8.2
azure-cognitiveservices-speech==1.23.0
azure-storage-blob==12.13.1
seaborn==0.12.0